Scripts/requirements_no_ui.txt
torch
numpy
scipy
protobuf
accelerate
sentencepiece
transformers
diffusers>=0.30.0
Pillow===10.4.0
pygltflib
rembg
opencv-python
onnxruntime
